+++ apps/mobile/src/components/signUp/SignupStep/EmailPassword/Email.tsx
@@ -7,8 +7,7 @@ import { EMAIL_MESSAGE } from '../../../../core/signUp/errorMessage';
 import { useUserEmail } from '../../../../hooks/queries/user';
 import { role } from '../../../../recoil/common/role';
 import { isNextStep } from '../../../../recoil/signUp/isNextStep';
-import { checkEmailForm } from '../../../../utils/signUp/checkForm';
-import { checkIsResend } from '../../../../utils/signUp/checkIsResendCode';
+import { checkEmailForm, checkIsResend } from '../../../../utils/signUp/check';
 import { EMAIL_RULE } from '../../../../validation/rules';
 import InputForm from '../../../common/Form/inputForm';
 import { StyledInput } from '../../../common/Input';
