--- apps/mobile/src/utils/signUp/checkEmailVerified.ts
@@ -1,5 +1,0 @@
-import { EMAIL_MESSAGE } from "../../core/signUp/errorMessage";
-
-export function checkEmailVerified(value: string | undefined) {
-  return value === EMAIL_MESSAGE.VERIFY;
-}
