--- apps/mobile/src/utils/signUp/checkEmailVerifyOKToSuccess.ts
@@ -1,5 +1,0 @@
-import { EMAIL_MESSAGE } from "../../core/signUp/errorMessage";
-
-export function checkEmailVerifyOKToSuccess(value: string | undefined) {
-  return value === EMAIL_MESSAGE.VERIFY;
-}
