--- apps/mobile/src/utils/signUp/checkInputEmpty.ts
@@ -1,3 +1,0 @@
-export function checkInputEmpty(value: string) {
-  return value === "";
-}
