--- apps/mobile/src/utils/signUp/checkIsResendCode.ts
@@ -1,5 +1,0 @@
-import { EMAIL_MESSAGE } from "../../core/signUp/errorMessage";
-
-export function checkIsResend(message: string | undefined) {
-  return message === EMAIL_MESSAGE.TIME;
-}
