--- apps/mobile/src/utils/signUp/checkPasswordMatch.ts
@@ -1,3 +1,0 @@
-export function checkPasswordMatch(password: string, passwordConfirm: string) {
-  return password === passwordConfirm;
-}
