--- apps/mobile/src/utils/signUp/showPassword.ts
@@ -1,3 +1,0 @@
-export function showPassword(isShow: boolean) {
-  return isShow ? "text" : "password";
-}
